toyy/requirements.txt
MechanicalSoup==1.3.0
selenium==4.20.0
python-dotenv==1.0.1
